fix(core): generalize structuredContent fix for all MCP servers
Previously, the transport-level fix that populates structuredContent from
JSON text content was only applied to Xcode's xcrun mcpbridge. This meant
that any other MCP server returning tools with outputSchema would fail
with error -32600 ('has an output schema but did not return structured
content').

This change renames XcodeMcpBridgeFixTransport to
StructuredContentFixTransport and applies it to all stdio MCP servers.
The fix is safe: it only triggers when structuredContent is missing,
silently ignores non-JSON content, and never overwrites existing
structuredContent.

Fixes #21053
